+++ src/test/kotlin/de/magynhard/crystal/parser/CrystalParserTest.kt
@@ -263,6 +263,10 @@ class CrystalParserTest : ParsingTestCase("", "cr", CrystalParserDefinition()) {
         doTest(true)
     }
 
+    fun testBareSplatCallArguments() {
+        doTest(true)
+    }
+
     fun testMacroInterpolatedCallee() {
         doTest(true)
     }